requirements.txt
Flask
Flask-SocketIO
python-socketio
pcapy-ng
dpkt
SQLAlchemy
psutil
eventlet